Separated class related to likelihood from framework.py.
