+++ metrics/__init__.py
@@ -10,3 +10,4 @@ __all__ = [
             'make_yy',
             'make_c_index'
             ]
+
